+++ libheif/linux_build_libs.py
@@ -79,6 +79,8 @@ def build_lib_linux(url: str, name: str, musl: bool = False):
                         "libheif/011-fix-loading-alpha-image.patch",
                         "libheif/012-fix-do-not-pad-16x16-AOM.patch",
                         "libheif/013-fix-enable-lossless-AOM.patch",
+                        "libheif/014-expose-aom-decoder-errors.patch",
+                        "libheif/015-aom-all-intra.patch",
                         "libheif/020-fix-wrong-copy-size.patch",
                     ):
                         patch_path = path.join(_linux_dir, patch)